requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.7.0
python-dotenv>=1.0.0
elasticsearch>=8.11.0,<9.0.0
requests>=2.31.0

pyjwt>=2.8.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0


